rename : member -> auth로 rename
