+++ app/src/main/java/org/android/go/sopt/auth/SignUpActivity.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.member
+package org.android.go.sopt.auth
 
 import android.os.Bundle
 import android.view.MotionEvent
