+++ app/src/main/java/org/android/go/sopt/auth/SignUpViewModel.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.member
+package org.android.go.sopt.auth
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MediatorLiveData
