+++ app/src/main/res/layout/activity_sign_in.xml
@@ -7,14 +7,14 @@
 
         <variable
             name="viewModel"
-            type="org.android.go.sopt.member.SignInViewModel" />
+            type="org.android.go.sopt.auth.SignInViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:paddingHorizontal="30dp"
-        tools:context=".member.SignInActivity">
+        tools:context=".auth.SignInActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
