+++ app/src/main/res/layout/activity_sign_up.xml
@@ -7,14 +7,14 @@
 
         <variable
             name="viewModel"
-            type="org.android.go.sopt.member.SignUpViewModel" />
+            type="org.android.go.sopt.auth.SignUpViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:paddingHorizontal="30dp"
-        tools:context=".member.SignUpActivity">
+        tools:context=".auth.SignUpActivity">
 
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
